# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the backend for SkillFort.",
        "title": "SkillFort-backend API docs",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:9000",
    "basePath": "/api/sf/v1",
    "paths": {
...
        "/user": {
            "get": {
                "description": "Get all active users",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Get all active users",
                "parameters": [
                    {
                        "type": "string",
                        "description": "search",
                        "name": "search",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responseModel.User"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.DataResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Saves an instance of User",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Save User",
                "parameters": [
                    {
                        "description": "user",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requestModel.User"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "501": {
                        "description": "Not Implemented",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/user/autocomplete": {
            "get": {
                "description": "Get  Users for AutoComplete",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Get  Users for AutoComplete",
                "parameters": [
                    {
                        "type": "string",
                        "description": "name",
                        "name": "name",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "role",
                        "name": "role",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entities.User"
                        }
                    }
                }
            }
        },
        "/user/channel": {
            "post": {
                "description": "Saves an instance of User Channel Detail",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "UserChannelDetail"
                ],
                "summary": "Save User Channel Detail",
                "parameters": [
                    {
                        "description": "userChannelDetail",
                        "name": "userChannelDetail",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/requestModel.UserChannelDetail"
                            }
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "501": {
                        "description": "Not Implemented",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/user/channel/accessible": {
            "get": {
                "description": "Gets the channel the user can access, to be used for channel switcher",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Gets the channel the user can access",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User id",
                        "name": "userId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/user/channel/{id}": {
            "put": {
                "description": "Update an instance of User Channel Detail",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "UserChannelDetail"
                ],
                "summary": "Update User Channel Detail",
                "parameters": [
                    {
                        "description": "userconfig",
                        "name": "userChannelDetail",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requestModel.UserChannelDetail"
                        }
                    },
                    {
                        "type": "integer",
                        "description": "userChannelDetailId",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "501": {
                        "description": "Not Implemented",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "delete": {
                "description": "Deletes an instance of User Channel Detail",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "UserChannelDetail"
                ],
                "summary": "Delete User Channel Detail",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "user channel detail id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/user/config": {
            "post": {
                "description": "Saves an instance of User Config",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "UserConfig"
                ],
                "summary": "Save User Config",
                "parameters": [
                    {
                        "description": "userconfig",
                        "name": "userConfig",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requestModel.UserConfig"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "501": {
                        "description": "Not Implemented",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/user/config/": {
            "get": {
                "description": "Get an instance of UserConfig",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "UserConfig"
                ],
                "summary": "Get a specific  UserConfig",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "userId",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entities.UserConfig"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.DataResponse"
                        }
                    }
                }
            }
        },
        "/user/config/{id}": {
            "put": {
                "description": "Update an instance of User Config",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "UserConfig"
                ],
                "summary": "Update User Config",
                "parameters": [
                    {
                        "description": "userconfig",
                        "name": "userConfig",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requestModel.UserConfig"
                        }
                    },
                    {
                        "type": "integer",
                        "description": "userconfigid",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "501": {
                        "description": "Not Implemented",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/user/forgot-password": {
            "post": {
                "description": "Reset User Password",
                "tags": [
                    "User"
                ],
                "summary": "Reset User Password",
                "parameters": [
                    {
                        "type": "string",
                        "description": "email",
                        "name": "email",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/user/login": {
            "post": {
                "description": "Saves an instance of User",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Login User",
                "parameters": [
                    {
                        "description": "login",
                        "name": "userLogin",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requestModel.Login"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
        "/user/reset-password": {
            "post": {
                "description": "Update User Password",
                "tags": [
                    "User"
                ],
                "summary": "Update User Password",
                "parameters": [
                    {
                        "type": "string",
                        "description": "resetString",
                        "name": "resetString",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "password",
                        "name": "password",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/user/switch-channel": {
            "get": {
                "description": "Refresh JWT Token and switches the user to the requested channel",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Refresh JWT Token and Switches the Channel",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Channel id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
...
    },
    "definitions": {
...
        "responseModel.User": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "extension": {
                    "type": "string"
                },
                "firstName": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "isActive": {
                    "type": "boolean"
                },
                "isLoggedIn": {
                    "type": "boolean"
                },
                "isLoginDisabled": {
                    "type": "boolean"
                },
                "lastLoginTime": {
                    "type": "string"
                },
                "lastName": {
                    "type": "string"
                },
                "loginFailureCounter": {
                    "type": "integer"
                },
                "phoneNumber": {
                    "type": "string"
                },
                "resetPasswordString": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "userChannelDetails": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/responseModel.UserChannelDetail"
                    }
                }
            }
        },
...
    }
}